refactor: dsl 용 Query 위치 이동
