+++ src/main/java/com/bob/core/application/post/PostQueryService.java
@@ -15,7 +15,6 @@ import org.springframework.transaction.annotation.Transactional;
 import com.bob.core.application.post.dto.query.ReadMemberPostsQuery;
 import com.bob.core.application.post.dto.query.ReadPostDetailQuery;
 import com.bob.core.application.post.dto.query.ReadPostFavoritesQuery;
-import com.bob.core.application.post.dto.query.ReadPostsQuery;
 import com.bob.core.application.post.dto.result.PostDetail;
 import com.bob.core.application.post.dto.result.PostSummaries;
 import com.bob.core.application.post.port.in.PostReader;
@@ -28,6 +27,7 @@ import com.bob.core.application.post.port.result.PostFile;
 import com.bob.core.application.post.port.result.PostMember;
 import com.bob.core.domain.post.Post;
 import com.bob.core.domain.post.repository.PostRepository;
+import com.bob.core.domain.post.repository.dsl.query.ReadPostsQuery;
 import com.bob.global.exception.exceptions.ApplicationException;
 import com.bob.global.exception.response.ApplicationError;
