+++ src/main/java/com/bob/core/application/post/port/in/PostReader.java
@@ -7,10 +7,10 @@ import org.springframework.data.domain.Pageable;
 import com.bob.core.application.post.dto.query.ReadMemberPostsQuery;
 import com.bob.core.application.post.dto.query.ReadPostDetailQuery;
 import com.bob.core.application.post.dto.query.ReadPostFavoritesQuery;
-import com.bob.core.application.post.dto.query.ReadPostsQuery;
 import com.bob.core.application.post.dto.result.PostDetail;
 import com.bob.core.application.post.dto.result.PostSummaries;
 import com.bob.core.domain.post.Post;
+import com.bob.core.domain.post.repository.dsl.query.ReadPostsQuery;
 
 public interface PostReader {
