+++ src/main/java/com/bob/core/application/trade/TradeQueryService.java
@@ -22,7 +22,6 @@ import com.bob.core.application.trade.dto.query.ReadPostTradeQuery;
 import com.bob.core.application.trade.dto.query.ReadPostTradesQuery;
 import com.bob.core.application.trade.dto.query.ReadTradeDetailQuery;
 import com.bob.core.application.trade.dto.query.ReadTradeStatusMapQuery;
-import com.bob.core.application.trade.dto.query.ReadTradesQuery;
 import com.bob.core.application.trade.dto.result.TradeDetail;
 import com.bob.core.application.trade.dto.result.TradeSummaries;
 import com.bob.core.application.trade.dto.result.TradeSummary;
@@ -38,6 +37,7 @@ import com.bob.core.application.trade.port.result.TradeMember;
 import com.bob.core.application.trade.port.result.TradePost;
 import com.bob.core.domain.trade.Trade;
 import com.bob.core.domain.trade.repository.TradeRepository;
+import com.bob.core.domain.trade.repository.dsl.query.ReadTradesQuery;
 import com.bob.global.exception.exceptions.ApplicationException;
 import com.bob.global.exception.response.ApplicationError;
