+++ src/test/java/com/bob/core/application/trade/port/in/TradeReaderTest.java
@@ -21,12 +21,12 @@ import com.bob.core.application.trade.dto.query.ReadPostTradeQuery;
 import com.bob.core.application.trade.dto.query.ReadPostTradesQuery;
 import com.bob.core.application.trade.dto.query.ReadTradeDetailQuery;
 import com.bob.core.application.trade.dto.query.ReadTradeStatusMapQuery;
-import com.bob.core.application.trade.dto.query.ReadTradesQuery;
-import com.bob.core.application.trade.dto.query.SearchKey;
 import com.bob.core.application.trade.dto.result.TradeDetail;
 import com.bob.core.application.trade.dto.result.TradeSummaries;
 import com.bob.core.application.trade.dto.result.internal.PostTrade;
 import com.bob.core.domain.trade.Trade;
+import com.bob.core.domain.trade.repository.dsl.query.ReadTradesQuery;
+import com.bob.core.domain.trade.repository.dsl.query.SearchKey;
 import com.bob.core.domain.trade.status.Status;
 import com.bob.global.exception.exceptions.ApplicationException;
 import com.bob.support.annotation.ContainerTest;
