+++ src/test/java/com/bob/support/fixture/post/dto/query/PostQueryFixture.java
@@ -3,10 +3,10 @@ package com.bob.support.fixture.post.dto.query;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.bob.core.application.post.dto.query.ReadPostsQuery;
-import com.bob.core.application.post.dto.query.condition.SearchKey;
-import com.bob.core.application.post.dto.query.condition.SearchPrice;
-import com.bob.core.application.post.dto.query.condition.SortKey;
+import com.bob.core.domain.post.repository.dsl.query.ReadPostsQuery;
+import com.bob.core.domain.post.repository.dsl.query.SearchKey;
+import com.bob.core.domain.post.repository.dsl.query.SearchPrice;
+import com.bob.core.domain.post.repository.dsl.query.SortKey;
 
 public class PostQueryFixture {
